src/requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
Django==4.0
django-dotenv==1.4.2
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.21.1
importlib-metadata==4.10.0
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
jsonschema==4.3.2
Markdown==3.3.6
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.5.2
pytest-dotenv==0.5.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
zipp==3.6.0
